init backend class
